Netflix Clone
This repository contains a clone of the popular streaming platform Netflix, built using HTML and CSS only. The project aims to replicate the user interface and basic functionality of Netflix, including features such as displaying featured content, listing movies and TV shows by categories, and implementing a responsive design for various devices.

Features:

Pure HTML and CSS for the structure and styling of the web pages.
Responsive design to ensure compatibility across different screen sizes and devices.
Mimics Netflix's color scheme, fonts, and layout to create an authentic user experience.
Organized structure with clear separation of components for easy maintenance and scalability.
Contributions:
Contributions to this project are welcome! Whether you're interested in adding new features, improving existing code, or fixing bugs, feel free to submit pull requests. Please adhere to the project's coding standards and guidelines.

Usage:
To run the Netflix clone locally, simply clone the repository and open the index.html file in your web browser.

Disclaimer:
This project is created for educational purposes and as a personal portfolio piece. It is not affiliated with or endorsed by Netflix. All trademarks and registered trademarks are the property of their respective owners.